+++ foliant/config/base.py
@@ -32,7 +32,7 @@ class BaseParser(object):
         :returns: Dictionary representing the YAML tree
         '''
 
-        self.logger.info('Parsing started.')
+        self.logger.info('Parsing started')
 
         with open(self.config_path, encoding='utf8') as config_file:
             config = {**self._defaults, **load(config_file, Loader)}
@@ -40,7 +40,7 @@ class BaseParser(object):
             config['src_dir'] = Path(config['src_dir']).expanduser()
             config['tmp_dir'] = Path(config['tmp_dir']).expanduser()
 
-            self.logger.info(f'Parsing completed.')
+            self.logger.info(f'Parsing completed')
 
             self.logger.debug(f'Config: {config}')